refactor: address PR review feedback
- Extract failWithFlow() helper to deduplicate observe/observeValues
- Add `import kotlinx.coroutines.flow.flow`, remove fully-qualified calls
- Add clarifying comment in read() for null config path
- Split FakePeripheralEnhancedTest into Write/Delay/ErrorInjection tests
- Remove duplicate tests (Turbine versions in FlowTest are canonical)
- Make BleException a data class for structural equality
- Fix ktlint violations in StubPeripheral and Lincheck tests
